kiwoom/requirements.txt
blinker==1.8.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.0.3
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
